requirements.txt
python-dotenv
python-telegram-bot
google-genai
